Splitted BookDAO into {BookDAO - BookInstanceDAO - BorrowingDAO} And added classes implementing those interfaces in dao/src/main/../
