+++ dao/src/main/java/hu/uni/miskolc/iit/softwaretesting/dao/BookDaoXMLImpl.java
@@ -1,7 +1,10 @@
 package hu.uni.miskolc.iit.softwaretesting.dao;
 
 import hu.uni.miskolc.iit.softwaretesting.exceptions.*;
-import hu.uni.miskolc.iit.softwaretesting.model.*;
+import hu.uni.miskolc.iit.softwaretesting.model.Book;
+import hu.uni.miskolc.iit.softwaretesting.model.Genre;
+import hu.uni.miskolc.iit.softwaretesting.model.Password;
+import hu.uni.miskolc.iit.softwaretesting.model.Reader;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.w3c.dom.Document;
@@ -9,6 +12,7 @@ import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
+
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
@@ -20,9 +24,6 @@ import javax.xml.transform.stream.StreamResult;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.*;
 
 /**
@@ -30,6 +31,22 @@ import java.util.*;
  */
 @Repository
 public class BookDaoXMLImpl implements BookDAO {
+
+    /*----------------------------------------------------------------------------------------------------------*/
+
+    //Creating a private static instance of our class to make it a Singleton class
+    private static BookDaoXMLImpl instance = new BookDaoXMLImpl();
+
+    //Making the constructor private to make it used only by this claas to make it a Singleton class
+    private BookDaoXMLImpl() {
+    }
+
+    //A getter of the one and only instance of this class (Singleton class)
+    public static BookDaoXMLImpl getInstance() {
+        return instance;
+    }
+
+    /*----------------------------------------------------------------------------------------------------------*/
 
     /**
      * The document to manipulate.
@@ -308,435 +325,6 @@ public class BookDaoXMLImpl implements BookDAO {
     }
 
     /**
-     * Creates a new book instance in database.
-     *
-     * @param bookInstance Data container of the new book instance.
-     */
-    public void createBookInstance(BookInstance bookInstance) throws AlreadyExistingBookInstanceException, PersistenceException {
-        if (this.isBookInstanceExists(bookInstance.getInventoryNumber())) {
-            throw new AlreadyExistingBookInstanceException();
-        }
-
-        try {
-            Element newBookInstance = document.createElement("bookInstance");
-
-            Element inventoryNumber = document.createElement("inventoryNumber");
-            inventoryNumber.appendChild(document.createTextNode(String.valueOf(bookInstance.getInventoryNumber())));
-            newBookInstance.appendChild(inventoryNumber);
-
-            Element bookISBN = document.createElement("bookISBN");
-            bookISBN.appendChild(document.createTextNode(String.valueOf(bookInstance.getBook().getIsbn())));
-            newBookInstance.appendChild(bookISBN);
-
-            Element isLoaned = document.createElement("isLoaned");
-            isLoaned.appendChild(document.createTextNode(Boolean.toString(bookInstance.isLoaned())));
-            newBookInstance.appendChild(isLoaned);
-
-            document.getElementsByTagName("bookInstances").item(0).appendChild(newBookInstance);
-            this.serializeDOM();
-        } catch (TransformerException|IOException e) {
-            throw new PersistenceException("Failed to create new book instance", e);
-        }
-    }
-
-    /**
-     * Returns all book instances from the database.
-     */
-    public ArrayList<BookInstance> getAllBookInstances() throws BookInstanceNotFoundException {
-        NodeList bookInstances = document.getElementsByTagName("bookInstance");
-        ArrayList<BookInstance> results = new ArrayList<>();
-
-        try {
-            for (int i=0; i < bookInstances.getLength(); i++) {
-                Element current = (Element) bookInstances.item(i);
-                results.add(
-                        new BookInstance(
-                                Long.valueOf(this.getNodeValue(current, "inventoryNumber")),
-                                this.getBookByISBN(Long.valueOf(this.getNodeValue(current, "bookISBN"))),
-                                Boolean.valueOf(this.getNodeValue(current, "isLoaned"))
-                        )
-                );
-            }
-        } catch (BookNotFoundException e) {
-            throw new BookInstanceNotFoundException(e);
-        }
-
-        return results;
-    }
-
-    /**
-     * Returns all instances of the given book.
-     *
-     * @param book The book object, which we search instances of.
-     */
-    public ArrayList<BookInstance> getAllInstancesOfBook(Book book) throws BookNotFoundException, BookInstanceNotFoundException {
-        NodeList bookInstances = document.getElementsByTagName("bookInstance");
-        String bookISBN = String.valueOf(book.getIsbn());
-        ArrayList<BookInstance> results = new ArrayList<>();
-
-        for (int i=0; i < bookInstances.getLength(); i++) {
-            Element current = (Element) bookInstances.item(i);
-            if (this.getNodeValue(current, "bookISBN").equals(bookISBN)) {
-                results.add(
-                        new BookInstance(
-                                Long.valueOf(this.getNodeValue(current, "inventoryNumber")),
-                                this.getBookByISBN(Long.valueOf(this.getNodeValue(current, "bookISBN"))),
-                                Boolean.valueOf(this.getNodeValue(current, "isLoaned"))
-                        )
-                );
-            }
-        }
-
-        if (results.isEmpty()) {
-            throw new BookInstanceNotFoundException();
-        }
-
-        return results;
-    }
-
-    /**
-     * Returns all borrowable instances of the given book.
-     *
-     * @param book The book object, which we search instances of.
-     */
-    public ArrayList<BookInstance> getAvailableInstancesOfBook(Book book) throws BookNotFoundException, BookInstanceNotFoundException {
-        NodeList bookInstances = document.getElementsByTagName("bookInstance");
-        String bookISBN = String.valueOf(book.getIsbn());
-        ArrayList<BookInstance> results = new ArrayList<>();
-
-        for (int i=0; i < bookInstances.getLength(); i++) {
-            Element current = (Element) bookInstances.item(i);
-
-            if (this.getNodeValue(current, "bookISBN").equals(bookISBN) &&
-                    !Boolean.valueOf(this.getNodeValue(current, "isLoaned"))) {
-                results.add(
-                        new BookInstance(
-                                Long.valueOf(this.getNodeValue(current, "inventoryNumber")),
-                                this.getBookByISBN(Long.valueOf(this.getNodeValue(current, "bookISBN"))),
-                                Boolean.valueOf(this.getNodeValue(current, "isLoaned"))
-                        )
-                );
-            }
-        }
-
-        if (results.isEmpty()) {
-            throw new BookInstanceNotFoundException();
-        }
-
-        return results;
-    }
-
-    /**
-     * Returns the book instance with the given inventoryNumber.
-     *
-     * @param inventoryNumber Inventory number of the book instance.
-     */
-    public BookInstance getBookInstanceByInventoryNumber(long inventoryNumber) throws BookInstanceNotFoundException, BookNotFoundException {
-        NodeList bookInstances = document.getElementsByTagName("bookInstance");
-
-        for (int i=0; i < bookInstances.getLength(); i++) {
-            Element current = (Element) bookInstances.item(i);
-            if (Long.valueOf(this.getNodeValue(current, "inventoryNumber")).equals(inventoryNumber)) {
-                return new BookInstance(
-                        Long.valueOf(this.getNodeValue(current, "inventoryNumber")),
-                        this.getBookByISBN(Long.valueOf(this.getNodeValue(current, "bookISBN"))),
-                        Boolean.valueOf(this.getNodeValue(current, "isLoaned"))
-                );
-            }
-        }
-
-        throw new BookInstanceNotFoundException();
-    }
-
-    /**
-     * Updates the book instance in the database.
-     *
-     * @param bookInstance Data container of the new book instance.
-     *                     InventoryNumber field must be set to the desired book instance's to update.
-     */
-    public void updateBookInstance(BookInstance bookInstance) throws BookInstanceNotFoundException, PersistenceException {
-        NodeList bookInstances = document.getElementsByTagName("bookInstance");
-
-        for (int i=0; i < bookInstances.getLength(); i++) {
-            Element current = (Element) bookInstances.item(i);
-            if (this.getNodeValue(current, "inventoryNumber").equals(String.valueOf(bookInstance.getInventoryNumber()))) {
-                this.getNode(current, "bookISBN").setTextContent(String.valueOf(bookInstance.getBook().getIsbn()));
-                this.getNode(current, "isLoaned").setTextContent(Boolean.toString(bookInstance.isLoaned()));
-            }
-        }
-
-        try {
-            this.serializeDOM();
-        } catch (TransformerException|IOException e) {
-            throw new PersistenceException(e);
-        }
-    }
-
-    /**
-     * Deletes the book instance from the database.
-     *
-     * @param bookInstance BookInstance object with the eraseable book instance's inventory number.
-     */
-    public void deleteBookInstance(BookInstance bookInstance) throws BookInstanceNotFoundException {
-        NodeList bookInstances = document.getElementsByTagName("bookInstance");
-        String inventoryNo = String.valueOf(bookInstance.getInventoryNumber());
-
-        for (int i=0; i < bookInstances.getLength(); i++) {
-            Element current = (Element) bookInstances.item(i);
-            if (this.getNodeValue(current, "inventoryNumber").equals(inventoryNo)) {
-                current.getParentNode().removeChild(current);
-            }
-        }
-    }
-
-    /**
-     * Creates a new borrowing in the database.
-     *
-     * @param borrowing Data container of the new borrowing.
-     */
-    public void createBorrowing(Borrowing borrowing) throws AlreadyExistingBorrowingException, PersistenceException {
-        if (this.isBorrowingExists(borrowing.getBorrowID())) {
-            throw new AlreadyExistingBorrowingException();
-        }
-
-        try {
-            Element newBorrowing = document.createElement("borrowing");
-
-            Element borrowID = document.createElement("borrowID");
-            borrowID.appendChild(document.createTextNode(String.valueOf(borrowing.getBorrowID())));
-            newBorrowing.appendChild(borrowID);
-
-            Element readerUsername = document.createElement("readerUsername");
-            readerUsername.appendChild(document.createTextNode(borrowing.getReader().getUsername()));
-            newBorrowing.appendChild(readerUsername);
-
-            DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-
-            Element creationDate = document.createElement("creationDate");
-            creationDate.appendChild(document.createTextNode(dateFormat.format(borrowing.getCreationDate())));
-            newBorrowing.appendChild(creationDate);
-
-            Element expirationDate = document.createElement("expirationDate");
-            expirationDate.appendChild(document.createTextNode(dateFormat.format(borrowing.getExpirationDate())));
-            newBorrowing.appendChild(expirationDate);
-
-            Element status = document.createElement("status");
-            status.appendChild(document.createTextNode(borrowing.getStatus().toString()));
-            newBorrowing.appendChild(status);
-
-            Element bookInstanceInventoryNo = document.createElement("bookInstanceInventoryNumber");
-            bookInstanceInventoryNo.appendChild(document.createTextNode(
-                    String.valueOf(borrowing.getBookInstance().getInventoryNumber())
-            ));
-            newBorrowing.appendChild(bookInstanceInventoryNo);
-
-            document.getElementsByTagName("borrowings").item(0).appendChild(newBorrowing);
-            this.serializeDOM();
-        } catch (TransformerException|IOException e) {
-            throw new PersistenceException(e);
-        }
-    }
-
-    /**
-     * Returns all of the borrowings from the database.
-     */
-    public ArrayList<Borrowing> getAllBorrowings() throws NoBorrowingsFoundException {
-        NodeList borrowings = document.getElementsByTagName("borrowing");
-        ArrayList<Borrowing> results = new ArrayList<>();
-
-        try {
-            for(int i=0; i < borrowings.getLength(); i++) {
-                Element current = (Element) borrowings.item(i);
-
-                DateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-
-                results.add(
-                        new Borrowing(
-                                Long.valueOf(this.getNodeValue(current, "borrowID")),
-                                this.getReaderByUsername(this.getNodeValue(current, "readerUsername")),
-                                format.parse(this.getNodeValue(current, "creationDate")),
-                                format.parse(this.getNodeValue(current, "expirationDate")),
-                                BorrowStatus.valueOf(this.getNodeValue(current, "status")),
-                                this.getBookInstanceByInventoryNumber(
-                                        Long.valueOf(this.getNodeValue(current, "bookInstanceInventoryNumber"))
-                                )
-                        )
-                );
-            }
-        } catch (ParseException| BookInstanceNotFoundException |BookNotFoundException|NotExistingReaderException e) {
-            throw new NoBorrowingsFoundException(e);
-        }
-
-        return results;
-    }
-
-    /**
-     * Returns all borrowings with the given status.
-     *
-     * @param status The searched borrowing status.
-     */
-    public Collection<Borrowing> getBorrowingsByStatus(BorrowStatus status) throws NotExistingBorrowingException {
-        NodeList borrowings = document.getElementsByTagName("borrowing");
-        ArrayList<Borrowing> results = new ArrayList<>();
-        String searchedStatus = status.toString();
-
-        try {
-            for (int i=0; i < borrowings.getLength(); i++) {
-                Element current = (Element) borrowings.item(i);
-                if (this.getNodeValue(current, "status").equals(searchedStatus)) {
-                    DateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-
-                    results.add(
-                            new Borrowing(
-                                    Long.valueOf(this.getNodeValue(current, "borrowID")),
-                                    this.getReaderByUsername(this.getNodeValue(current, "readerUsername")),
-                                    format.parse(this.getNodeValue(current, "creationDate")),
-                                    format.parse(this.getNodeValue(current, "expirationDate")),
-                                    BorrowStatus.valueOf(this.getNodeValue(current, "status")),
-                                    this.getBookInstanceByInventoryNumber(
-                                            Long.valueOf(this.getNodeValue(current, "bookInstanceInventoryNumber"))
-                                    )
-                            )
-                    );
-                }
-            }
-        } catch (ParseException| BookInstanceNotFoundException |BookNotFoundException|NotExistingReaderException e) {
-            throw new NotExistingBorrowingException(e);
-        }
-
-        return results;
-    }
-
-    /**
-     * Returns all borrowings of the given reader.
-     *
-     * @param reader The searched reader.
-     */
-    public Collection<Borrowing> getBorrowingsOfReader(Reader reader) throws NotExistingBorrowingException, NotExistingReaderException {
-        NodeList borrowings = document.getElementsByTagName("borrowing");
-        ArrayList<Borrowing> results = new ArrayList<>();
-        String searchedReaderUsername = reader.getUsername();
-        this.getReaderByUsername(searchedReaderUsername);
-
-        try {
-            for (int i=0; i < borrowings.getLength(); i++) {
-                Element current = (Element) borrowings.item(i);
-                if (this.getNodeValue(current, "readerUsername").equals(searchedReaderUsername)) {
-                    DateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-
-                    results.add(
-                            new Borrowing(
-                                    Long.valueOf(this.getNodeValue(current, "borrowID")),
-                                    this.getReaderByUsername(this.getNodeValue(current, "readerUsername")),
-                                    format.parse(this.getNodeValue(current, "creationDate")),
-                                    format.parse(this.getNodeValue(current, "expirationDate")),
-                                    BorrowStatus.valueOf(this.getNodeValue(current, "status")),
-                                    this.getBookInstanceByInventoryNumber(
-                                            Long.valueOf(this.getNodeValue(current, "bookInstanceInventoryNumber"))
-                                    )
-                            )
-                    );
-                }
-            }
-
-            if (results.isEmpty()) {
-                throw new NotExistingBorrowingException();
-            }
-        } catch (ParseException| BookInstanceNotFoundException |BookNotFoundException e) {
-            throw new NotExistingBorrowingException(e);
-        }
-
-        return results;
-    }
-
-    /**
-     * Returns the borrowing with the given id.
-     * @param id ID to search.
-     * @return Borrowing with the given id.
-     * @throws NotExistingBorrowingException
-     */
-    public Borrowing getBorrowingById(long id) throws NotExistingBorrowingException {
-        NodeList borrowings = document.getElementsByTagName("borrowing");
-        Borrowing result = null;
-
-        try {
-
-            for (int i=0; i < borrowings.getLength(); i++) {
-                Element current = (Element) borrowings.item(i);
-                if (this.getNodeValue(current, "borrowID").equalsIgnoreCase(String.valueOf(id))) {
-                    DateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-                    result = new Borrowing(
-                            Long.valueOf(this.getNodeValue(current, "borrowID")),
-                            this.getReaderByUsername(this.getNodeValue(current, "readerUsername")),
-                            format.parse(this.getNodeValue(current, "creationDate")),
-                            format.parse(this.getNodeValue(current, "expirationDate")),
-                            BorrowStatus.valueOf(this.getNodeValue(current, "status")),
-                            this.getBookInstanceByInventoryNumber(
-                                    Long.valueOf(this.getNodeValue(current, "bookInstanceInventoryNumber"))
-                            )
-                    );
-                }
-            }
-        } catch (ParseException| BookInstanceNotFoundException |BookNotFoundException|NotExistingReaderException e) {
-            throw new NotExistingBorrowingException(e);
-        }
-
-        if (result == null)
-            throw new NotExistingBorrowingException();
-
-        return result;
-    }
-
-    /**
-     * Updates borrowing in the database.
-     *
-     * @param borrowing The borrowing object to update.
-     *                  BorrowID field have to contain the unique id of the borrowing to update.
-     */
-    public void updateBorrowing(Borrowing borrowing) throws NotExistingBorrowingException, PersistenceException {
-        NodeList borrowings = document.getElementsByTagName("borrowing");
-
-        for (int i=0; i < borrowings.getLength(); i++) {
-            Element current = (Element) borrowings.item(i);
-            if (this.getNodeValue(current, "borrowID").equals(String.valueOf(borrowing.getBorrowID()))) {
-                DateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-
-                this.getNode(current, "readerUsername").setTextContent(borrowing.getReader().getUsername());
-                this.getNode(current, "creationDate").setTextContent(format.format(borrowing.getCreationDate()));
-                this.getNode(current, "expirationDate").setTextContent(format.format(borrowing.getExpirationDate()));
-                this.getNode(current, "status").setTextContent(borrowing.getStatus().toString());
-            }
-        }
-
-        try {
-            this.serializeDOM();
-        } catch (TransformerException|IOException e) {
-            throw new PersistenceException(e);
-        }
-    }
-
-    /**
-     * Deletes the given borrowing from the database.
-     *
-     * @param borrowing The borrowing object to delete.
-     *                  BorrowID field have to contain the unique id of the borrowing to delete.
-     */
-    public void deleteBorrowing(Borrowing borrowing) throws NotExistingBorrowingException {
-        NodeList borrowings = document.getElementsByTagName("borrowing");
-        String borrowID = String.valueOf(borrowing.getBorrowID());
-
-        for (int i=0; i < borrowings.getLength(); i++) {
-            Element current = (Element) borrowings.item(i);
-            if (this.getNodeValue(current, "borrowID").equals(borrowID)) {
-                current.getParentNode().removeChild(current);
-                return;
-            }
-        }
-
-        throw new NotExistingBorrowingException();
-    }
-
-    /**
      * Returns the books with the given genre.
      *
      * @param genre The genre of the book.
@@ -861,7 +449,7 @@ public class BookDaoXMLImpl implements BookDAO {
      * @return
      * @throws NotExistingReaderException
      */
-    private Reader getReaderByUsername (String username) throws NotExistingReaderException {
+    public Reader getReaderByUsername (String username) throws NotExistingReaderException {
         NodeList users = document.getElementsByTagName("user");
 
         for (int i=0; i < users.getLength(); i++) {
@@ -886,7 +474,7 @@ public class BookDaoXMLImpl implements BookDAO {
      * @param node Root node.
      * @param nodeName The searched node's name.
      */
-    private Node getNode (Node node, String nodeName) {
+    public Node getNode (Node node, String nodeName) {
         NodeList nodes = node.getChildNodes();
         for (int i=0; i < nodes.getLength(); i++) {
             Node current = nodes.item(i);
@@ -903,7 +491,7 @@ public class BookDaoXMLImpl implements BookDAO {
      * @param node Root node
      * @param nodeName The searched node's name
      */
-    private String getNodeValue (Node node, String nodeName) {
+    public String getNodeValue (Node node, String nodeName) {
         return getNode(node, nodeName).getTextContent();
     }
 
@@ -911,34 +499,10 @@ public class BookDaoXMLImpl implements BookDAO {
      * Checks if book exists with the given ISBN number.
      * @param isbn ISBN to search for.
      */
-    private boolean isBookExists (long isbn) {
+    public boolean isBookExists (long isbn) {
         try {
             return this.getBookByISBN(isbn) != null;
         } catch (BookNotFoundException e) {
-            return false;
-        }
-    }
-
-    /**
-     * Checks if book instance exists with the given ISBN number.
-     * @param inventoryNo Inventory no. to search for.
-     */
-    private boolean isBookInstanceExists (long inventoryNo) {
-        try {
-            return this.getBookInstanceByInventoryNumber(inventoryNo) != null;
-        } catch (BookNotFoundException| BookInstanceNotFoundException e) {
-            return false;
-        }
-    }
-
-    /**
-     * Checks if borrowing exists with the given ISBN number.
-     * @param borrowID ID to search for.
-     */
-    private boolean isBorrowingExists (long borrowID) {
-        try {
-            return this.getBorrowingById(borrowID) != null;
-        } catch (NotExistingBorrowingException e) {
             return false;
         }
     }
@@ -948,7 +512,7 @@ public class BookDaoXMLImpl implements BookDAO {
      * @throws TransformerException
      * @throws IOException
      */
-    private void serializeDOM () throws TransformerException, IOException {
+    public void serializeDOM () throws TransformerException, IOException {
         DOMSource source = new DOMSource(this.document);
         FileWriter writer = new FileWriter(this.outputFile);
         StreamResult result = new StreamResult(writer);
